Refactor email reply endpoint to support threading and simplify request handling
- Updated the email reply route to support both email IDs and thread IDs, allowing replies to the latest message in a thread.
- Introduced a new request interface for simplified reply handling, including a replyAll option to include original CC recipients.
- Enhanced validation and logging for ID resolution and recipient determination.
- Removed legacy code and streamlined the email sending process, ensuring proper threading headers are included.
- Improved error handling and response structure for better clarity in API responses.

These changes enhance the functionality and user experience of the email reply system, making it more robust and user-friendly.
